Collapse views together for spaces
