+++ examples/apps/spaces/src/fluid-object/spacesView.tsx
@@ -5,45 +5,44 @@
 
 import React from "react";
 import "react-grid-layout/css/styles.css";
-import { ISpacesItemEntry } from "./spacesItemMap";
-import { ISpacesStorage, SpacesStorageView } from "./storage";
+import { spacesItemMap, templateDefinitions } from "./spacesItemMap";
+import { SpacesStorageView } from "./storage";
+import { Spaces } from "./spaces";
 import { SpacesToolbar } from "./spacesToolbar";
-import { ISpacesItem } from "./index";
 
 interface ISpacesViewProps {
-    itemMap: Map<string, ISpacesItemEntry>;
-    storage: ISpacesStorage<ISpacesItem>;
-    addItem(type: string): void;
-    getViewForItem: (item: ISpacesItem) => Promise<JSX.Element | undefined>;
-    getUrlForItem: (itemId: string) => string;
-    templates?: string[];
-    applyTemplate?(template: string): void;
+    model: Spaces;
 }
 
-/**
- * SpacesView is the full view of the Spaces component, including its toolbar and contained items.
- */
-export const SpacesView: React.FC<ISpacesViewProps> =
-    (props: React.PropsWithChildren<ISpacesViewProps>) => {
-        // Editable is a view-only concept; SpacesView is the authority.
-        const [editable, setEditable] = React.useState<boolean>(props.storage.itemList.size === 0);
+export const SpacesView: React.FC<ISpacesViewProps> = (props: ISpacesViewProps) => {
+    const { model } = props;
+    const [baseUrl, setBaseUrl] = React.useState<string | undefined>("");
+    const [editable, setEditable] = React.useState<boolean>(model.storageComponent.itemList.size === 0);
+    React.useEffect(() => {
+        const getBaseUrl = async () => {
+            setBaseUrl(await model.getBaseUrl());
+        };
 
-        return (
-            <div className="spaces-view">
-                <SpacesToolbar
-                    editable={editable}
-                    setEditable={setEditable}
-                    itemMap={props.itemMap}
-                    addItem={props.addItem}
-                    templates={props.templates}
-                    applyTemplate={props.applyTemplate}
-                />
-                <SpacesStorageView
-                    getViewForItem={props.getViewForItem}
-                    getUrlForItem={props.getUrlForItem}
-                    storage={props.storage}
-                    editable={editable}
-                />
-            </div>
-        );
-    };
+        getBaseUrl().catch((error) => {
+            console.error(error);
+        });
+    });
+    return (
+        <div className="spaces-view">
+            <SpacesToolbar
+                editable={editable}
+                setEditable={setEditable}
+                itemMap={spacesItemMap}
+                addItem={model.addItem}
+                templates={[...Object.keys(templateDefinitions)]}
+                applyTemplate={model.applyTemplate}
+            />
+            <SpacesStorageView
+                getViewForItem={model.getViewForItem}
+                getUrlForItem={(itemId: string) => `#${baseUrl}/${itemId}`}
+                storage={model.storageComponent}
+                editable={editable}
+            />
+        </div>
+    );
+};
